rename: 파일명 통일을 위해 user → firebaseUser로 변경
